2018.08.08: * fixed: trigonometric range fix before call to acos/asin functions
2018.08.06:
* fixed: minor compilation warnings


git-svn-id: svn+ssh://svn.code.sf.net/p/orbittools/orbittools/trunk@7 3e713a50-9f52-49c3-8e70-6593cb2f10db

+++ core/orbitTools/Defines.h
@@ -20,8 +20,8 @@ namespace qd
 #elif defined(ORBIT_TOOLS_ENABLE_QD_DD_INTEGRATION)
 namespace qd
 {
-    static inline dd_real to_real(int i) { return i; }
-    static inline dd_real to_real(int64_t i) { return i; }
+    static inline dd_real to_real(int i) { return double(i); }
+    static inline dd_real to_real(int64_t i) { return double(i); }
     static inline dd_real to_real(double d) { return d; }
 }
 
@@ -35,8 +35,8 @@ static inline double to_double(double d) { return d; }
 
 namespace qd
 {
-    static inline double to_real(int i) { return i; }
-    static inline double to_real(int64_t i) { return i; }
+    static inline double to_real(int i) { return double(i); }
+    static inline double to_real(int64_t i) { return double(i); }
     static inline double to_real(double d) { return d; }
 }
